Use temp. directory to hold results
Some system environments cause the fetch modle to fail with a checksum
mismatch error if certain characters are present in the destination
directory name.  This only affects the fetch module, and doesn't seem
linked to the ansible version.  To work around this, a temporary
directory is used to hold the results, before moving to the playbook directory.

Signed-off-by: Christopher Evich <cevich@redhat.com>