Testing independence from Jenkins + any cloud
Previously, using ADEPT by itself was possible, but not easy.  A lot of
tricks and setup were required before it could do much useful testing.
Much of that setup work was handled (internally) by jenkins and bunches
of scripts.  This makes for a very difficult situation when more
public-facing testing and ease-of-use both need to happen.

This change moves nearly all the previously private scripts/jobs,
under the control of a single, state-aware python script.  This
script is fed with simply formatted yaml instructions.  These
are all that's needed to completely bootstrap a CI-like Docker Autotest
run.

This allows actual production-testing setup data to be stored in-repo
and remain agnostic WRT actual tooling and/cloud used.  The only thing
that varies from one environment to another are the cloud-type and
cloud-specific credentials / security settings.

Finally, it allows additional non-public testing to be done in a uniform
way, since the execution and cloud layers are independent, the job
becomes a simple alteration of input variables.

Signed-off-by: Chris Evich <cevich@redhat.com>