Configure default autotested role to run a CI-set
There will be a CI-job dedicated to validating ADEPT and the CI
infrastructure itself.  For all other cases, these defaults will
be overridden, and the environments will be customized in different
ways.  Select subtests by exclusion, such that new content is
automaticly validated as it's added.  Exclude tests that are very
specific and/or intended to be highly customized for a particular setup.

Signed-off-by: Chris Evich <cevich@redhat.com>