Sync: Remote overwrite local, local updates remote
When rsyncing files between the local system and the kommandir, don't
overwrite kommandir files that are newer than local.  However,
when there are differences syncing back, always overwrite local
files with the remote files.  Since most of the important 'action'
is happening from the kommandir, the local node should reflect it's
files, not overwrite with it's own.  That will cause an inconsistent
state and difficult to debug situation.

Signed-off-by: Chris Evich <cevich@redhat.com>