Fix Autotest/Docker Autotest cache clash
The autotested role first transfers a copy of the autotest repository to
the test VM.  If recursive submodules were used to clone that repo,
it will include a (likely very) old version of docker autotest.  Later,
when the cached (modern) docker-autotest is also transfered, it will end
up with bits that have long since been removed.

Fix this by first removing the docker autotest path before doing the
cache copy.

Signed-off-by: Chris Evich <cevich@redhat.com>