Devel (#59)
Fix installed role when installed_rpms contains urls

When specifying a mix of rpm URLs and package names, the former cache state can pollute yum module operations (i.e. if repos have changed on disk, but have the same name) and prior yum commands have run.  Since it's difficult to know the prior state of cache, this change forces clearing the yum cache every time the role is applied.  This will slow down operations considerably, however state will be consistent.

Ansible docs specify the ``state`` attribute value ``present`` == ``installed`` (for the ``yum`` module).  In practice this is not the case (Ansible 2.1).  Rather the ``installed`` and ``present`` behave differently when a package is specified by URL and an older package exists on the system.  In this case, when using ``present``, only the package name is checked.  With ``installed`` both the name and version are checked, and the newer one is used if available (from any source).  This change alters the default from ``present`` to ``installed`` (the ``state`` variable remains available to override this).

Signed-off-by: Chris Evich <cevich@redhat.com>