Add internal bypass for installed/updated
Normally when an ansible variable is overridden on the command-line, it
basically becomes read-only throughout, with all other role/task changes
being discarded (silently).  This is a problem for any roles or include
tasks which rely on passing in different values to suite their purposes.

In this case, the "updated" role, stuffs the 'installed_rpms' variable,
then includes the "installed" role, with state="updated".  However, if
installed_rpms was overridden, then incorrect behavior ensues.  There is
a similar problem with the lvm role.

This change opens up internal-role access to the "_installed_rpms"
variable, documented as such, not to be used outside of the playbook.
When defined and non-empty, it's value will be used in place of whatever
"installed_rpms" is set to.

Signed-off-by: Chris Evich <cevich@redhat.com>