Fix passing GUUID as name to alt. results2junit
Test VMs (peons) are all assigned a globally unique name to prevent
clashes with concurrently running jobs.  This is not a good name to use
for tagging test results because they will never align from run-to-run.

This change makes the peon and cloud metadata available to them via
host_vars.  Later, when the autotest role is applied, it's possible to
extract the peon's uncompounded (basic definition) name.  This is
guaranteed unique within a run, but will be (mostly) shared across
them.

Signed-off-by: Chris Evich <cevich@redhat.com>