Fix: Fail task not re-applied after 'always' block.
Instead of confirming the exit code is 0 or 77 w/in the block, move it
outside.  This allows the 'always' section to apply before flagging the
host as failed.  The only other solution is to set ``any_errors_fatal``
at the play level, which could impact other roles in unknown ways.

Signed-off-by: Chris Evich <cevich@redhat.com>