Cook results if autotest or results2junit fail
In order for results to always be collected, the exit
status of both autotest and the result conversion scripts
must be ignored.  Otherwise if the autotested role fails
on a host, the results directory will never get synchronized
back to the slave node.

However, it's equally bad if empty results are reported
because that can sometimes mean "everything was OK".  Add
detection for a missing results file, and drop some "cooked"
results in its place.

Also, update comments in the affected roles as to why
everything ignores errors.

Signed-off-by: Chris Evich <cevich@redhat.com>