Fix missing variable reference
It's possible the configuration for the autotested role is evaluated at
a different time that the tasks execute.  This can result in the
autotest_docker_path variable not being defined.

Move establishment of this value from w/in the playbook tasks to the
configuration defaults so it's always available and obvious.

Signed-off-by: Chris Evich <cevich@redhat.com>