Fix docker-latest peons use wrong d-s-s sysconfig
When /etc/sysconfig/docker points to 'docker-latest' then a different
version of d-s-s is used, which checks
/etc/sysconfig/docker-latest-storage-setup.  When configuring d-s-s on
peons which are to use docker-latest, the wrong file was being updated.
This caused docker-latest's d-s-s to fail (silently) and revert to using
loopback storage in some cases.  Fix this by adding a default
destination, then overriding it for docker-latest group members.

Signed-off-by: Chris Evich <cevich@redhat.com>