Fix latest fedora images missing python
There are a few Docker Autotest subtests which depend on python.  For
now, just use the default CentOS image.  This could lead to problems
later on if/when the kernel differences grow so large that the Fedora
Host Kernel can't support the CentOS container's userspace.

A more complete / proper fix for this is to stick a dockerfile somewhere
available by URL.  Pass the url to the file in the
``docker_autotest.test_image.build_url`` variable, and the autotested
role will take care of building the image + setting it up for testing.
A FIXME comment has been added to this effect.

Signed-off-by: Chris Evich <cevich@redhat.com>