Resolve TODO: Make autotested role context-aware
Nearly every other role in ADEPT runs a task-list that's conditional on
the specific context adept.py has passed in (setup, run, cleanup).  This
is a very useful scheme because it allows inspecting the
setup/configuration results prior to the start of testing.

This way, if any system goes completely belly-up because of a test,
additional debugging details will be preserved for later analysis.  It
also permits ADEPT-external steps to take place between setup/run. e.g.
updating CI status to a PR.

This commit extends the feature to the autotested role.

Signed-off-by: Chris Evich <cevich@redhat.com>