Restore feature: Copy docker-autotest related files
This was removed long ago due to dis-use and because it was too hard to
update/modify the copy or template list w/o having to redefine the
entire docker_autotest dict.  This change brings back in the copy
feature, and relocates it and the template feature into an included file
for modularity.

Signed-off-by: Chris Evich <cevich@redhat.com>