Fix clearing of config_custom
I don't remember why this was originally added.  However, it actively
prevents cache-stuffing, which is a useful feature in some cases.
Meaning, pre-populating / manipulating the repo. source contents before
it gets deployed it to the test VMs.

Support this use-case by not wiping the config_Custom directory when
transferring docker autotest from the cache to the VMs.

Signed-off-by: Chris Evich <cevich@redhat.com>