Allow empty default env. vars but not loaded
It's desireable for adept.py to fail with an error if a variable is
requested to be set from the parent environment, but doesn't exist or is
empty.  This is used to signal that a required environment variable was
not set.

However, there are occasions where an optional environment variable
should be passed through if it exists, otherwise an empty value
stored in it's place.  While adept.py supported storing a non-empty
default value (for a missing/empty parent env. var), it did not allow
and empty default through.

Add a private class and instance variable for storing the default, and process
it separately from any loaded value from the parent environment.

Signed-off-by: Chris Evich <cevich@redhat.com>