Fix pull tests trying to pull non-existent images
If test_image_fqin was produced by
``docker_autotest.test_image.build_url`` then it won't be available for
use in these tests.  Fix this by hard-coding a pair of public images
selected by the ansible_distribution magic variable.  This is not a
perfect fix, since it doesn't support testing non-Fedora/RHEL distros.
If that becomes a problem, some sort of lookup table should be used.

Signed-off-by: Chris Evich <cevich@redhat.com>