Use command not module for Bug workaround
At some point there were one or more bugs affecting the atomic command.
The workaround for them was to put selinux into permissive mode.
However, it was found the ansible selinux module modifies
/etc/sysconfig/selinux AND does a setenforce 0.

A subsequent atomic host upgrade will cause the to-be-deployed
os-tree (on reboot) to retain the 'permissive' value, even if it
is restored to 'enforcing' after the upgrade.  This doesn't
happen on non-ostree based hosts, because they don't "freeze" state at
the time of upgrade.

Fix all cases of atomic hosts selinux getting stuck in permissive
by just using the 'setenforce 0' command so enforcing mode is always
re-enabled upon reboot.

Signed-off-by: Chris Evich <cevich@redhat.com>

fixup! Use command not module for BZ1290281 workaround