Fix docker_test_images/puller
On ``fedora`` and ``fedoraatomic`` peons, we build a custom test image
in order to include some critical items missing from the base image (to
keep it's size small).  However, the default behavior of
docker_test_images/puller is to retrieve a CentOS image along with
whatever the default ``docker_registry_host``, ``docker_registry_user``,
``docker_repo_name``, and ``docker_repo_tag`` referr to.  The CentOS
image is not needed by any testing on Fedora, and because the default
test image is built, it's impossible to 'pull' it from anywhere.

Fix this by making sure the ``docker_test_images/puller`` is excluded in
``control.ini``, and just to make doubly sure, disable the sub-subtest
in ``subtests.ini`` by setting ``extra_fqins_csv`` empty.  Also
configure the ``builder`` sub-subtest with the same options used in
fedora/fedoratomic's ``autotested`` variable.  This ensures that on
these platforms, the image is always built and never pulled.

Signed-off-by: Chris Evich <cevich@redhat.com>