Fix puller 'referenced in control' error
Sub-subtests get mad if they're referenced for exclusion in control, but
not mentioned in the parent's subsubtests option.  Fixing this properly
is extraordinarily challenging because the control file doesn't know of
the subsubtests existence.  By the time it's known by the subtest
caller, it's too late, so it throws an error.  Take the easy road by
adding it back into the config. template, but keep it in the control.ini
exclude line.

Signed-off-by: Chris Evich <cevich@redhat.com>