Don't pass transition variables as ansible extra-vars
Both --extra-vars and -e, on the command-line or hauled
in from ``@filename`` have the unfortunate side-effect
of forcing those keys to be silently read-only for the
entire play.  Though it was heavily relied on previously,
it is no longer.  Instead, simple environment variables
should be used to make necessary adjustments under the
control of ``command`` action-items.  Playbooks can
then e.g. use ``var_files`` to bring those variables
in w/o the side-effects.

Signed-off-by: Chris Evich <cevich@redhat.com>