Fix single-source peon group list
If all groups to which a peon should belong, are encoded under a single
list, it makes list re-use difficult.  For example, if I want to define a
new peon based on an existing one.  I'd have to copy the entire thing,
even if all I want to do is change one group.

This change makes it so that any facts defined for a peon, whatever the source,
if the name matches the regex "*._groups", it will be used as a list of groups to join.
Finally, since variables are (in part) sourced from groups, this allows the group-addition list
search to be somewhat recursive.

To help relieve some of the brain-stress, also dump the latest
kommandir's facts into a ``results`` file.

Signed-off-by: Chris Evich <cevich@redhat.com>