Fix use of namespace for fedora_test_image
If it's there, then docker feels compelled to try and contact the
localhost registry (which doesn't exist). Instead, if the
default test image was deleted, we want tests to fail as early as
possible.

Also updated documentation. This change needs to be paired with
a similar change being made in Docker Autotest

Signed-off-by: Chris Evich <cevich@redhat.com>