Easy sub/subtest specifier for autotested role
When only a simple list of tests should run,
it was required to copy/modify the default
``control.ini.j2`` template.  Since this can
be combersome, add an optional ``job_subthings``
variable to the default template.  When non-empty
this provides a very easy way to specify a small
set of tests ro tun.  Also update documentation.

Signed-off-by: Chris Evich <cevich@redhat.com>