Fix 400 response attaching non-available volume
It's an error to attempt volume attachment before it's finished being
created (in OpenStack).  This problem wasn't noticed because volumes
were always created fast enough previously.  Fix this by examining the
volume status, and not attempting attachment until it's "available".

Also fix a bug when volume attachment fails, ``destroy()`` would leak
it.  This happens because it only deleted previously attached volumes.
Fix this by searching all volumes, and matching the server-id in it's
name - volumes always created this way.  If matched, include it in the
to-be-deleted list.

Lastly, do not kill the ``previous_responses`` cache upon
``OpenstackREST()`` init.  This makes debugging awful difficult since
the default behavior is to try and delete partially created resources.

Signed-off-by: Chris Evich <cevich@redhat.com>