Disable fact-caching + SSH/CPU optimize
This is used to save some time discovering / setting variables when
playbooks are run multiple times.  However, it's impossible to know if
the cache is valid or not and some facts always need to be set up at runtime.
Disabling the cache forces this to happen.  It also causes missing variables
to be actually missing - assisting with debugging.

Also add a few tweaks to improve CPU and SSH performance

Signed-off-by: Chris Evich <cevich@redhat.com>