Fix exekutir -> kommandir sync of .venv directory
Update the virtual environment directory name.

Update the README template name

This directory is used to contain runtime dependencies in the exekutir's
workspace.  It is seeded in the kommandir workspace in the event
a *nocloud* kommandir is being used - saving download time.  However,
with a remote kommandir, local system state cannot be predicted
therefor this directory should not be synchronized.

Signed-off-by: Chris Evich <cevich@redhat.com>