Fix: Kommandir disk space
By default, openstack provisioning does not grow the root filesystem.
This is done to allow easier customizing of the layout from w/in
ansible.  However, for the kommandir, all available disk space is
required for jobs.  Drop a custom userdata file, and reference it
from the openstack group variable file.

Signed-off-by: Chris Evich <cevich@redhat.com>