Fix kommandir PB ignore $ANSIBLE_PRIVATE_KEY_FILE
In some cases, a playbook running on the kommandir may want to use an
alternate private key file.  This is supported by setting the expected
variable in job.xn, for the ansible-playbook command.  However in the
``common`` role, the ``ansible_private_key_file`` fact hard-codes
the value instead of looking it up from the environment.  Fix this.

Signed-off-by: Chris Evich <cevich@redhat.com>