Fix unable to override git_cache_args from exekutir
When setting up a job's workspace on the kommandir VM, the
git_cache_args fact is consulted for the set of git repositories to
clone.  Later playbooks are then able to utilize those clones when
setting up peons.  However, because host_vars take precedence,
it is impossible to override this set of repos. to tailor for
a specific job.  Fix this by moving the default docker-autotest-based
default definition into the relevant role's defaults file.  This
places it at the lowest level of scoping possible, making it easy
to override as needed.

Signed-off-by: Chris Evich <cevich@redhat.com>