Gut kommandir roles/groups to be more generic
Previously ADEPT was solely dedicated to docker-autotest orchestration.
This is changing bit by bit, and having systems setup for that task
by default is getting in the way.  Remove all roles/setup from default
playbooks related to docker and autotest.  Update included host
definitions as well to be as minimal as possible.

Signed-off-by: Chris Evich <cevich@redhat.com>