Fix running setup more than once
ADEPT's setup should behave idempotently, meaning contexts can be
applied more than once.  A bug in the exekutir_workspace_setup role
relating to pre-existing results was preventing this.  Also,
the default openstack provisioning command was configured to
fail if a peon already existed.  Change this to be more tolerant
of existing peons.

Signed-off-by: Chris Evich <cevich@redhat.com>