Greatly simplify peon_up role (#203)
This role must not alter the state of the peon it's being used against -
such as marking it unreachable to ansible.  Previously this was
accomplished by executing a command from a template, but was horribly
complex.  Instead, simply run the 'ansible' command, using the 'ping'
module to check the peon.

Signed-off-by: Chris Evich <cevich@redhat.com>