Fix per-task become usage
Ansible allows specifying become/become_user on a per-task basis, but
only if it's not overriden by a connection variable.  Unf. this is
exactly what's done by default in ``kommandir/bin/adept_openstack.py``.
Remove this behavior, and instead rely on the default settings in
ansible.conf.  This allows plays/tasks to override them when/as needed.

Signed-off-by: Chris Evich <cevich@redhat.com>