Lower error -> warning for parsing message
One situation caught by the 'invalid' launched-at conditional is when a
server is first being created.  i.e. If a new VM is being created exactly
when ``server_expires_at()`` is called.  Instead of treating this as an
error, lower to a warning, and attempt to include the server name.
Otherwise fall-back to giving it's uuid in the message.

Add a unittest to verify this condition does not destroy a VM creation
in-flight.

Signed-off-by: Chris Evich <cevich@redhat.com>