Add new verbose logging level filter
This actually implements --verbose, by defining two levels
of INFO messages:

    logging.info("This is always displayed")
    logging.info(">Shown only if --verbose")

That is: if message begins with '>', it is filtered out
by default but shown if --verbose (or --debug). The
leading '>' is stripped.

I considered reversing the sense of the test: filtering
out INFO messages by default, and showing only those
starting with '!' or '*' or something. I think that
would be counterintuitive, so I went with show-unless->
even though that makes this diff larger.

Signed-off-by: Ed Santiago <santiago@redhat.com>