Fix missing remote ssh directory
Certain files/directories are excluded from synchronization of the
kommandir's workspace.  Previously the 'ssh' directory was added to this
list and should not have been.  The intention was to prevent jobs from
overwriting the 'ssh' directory for security-reasons.  Fix this in the
proper-places by ammending the exclusion when needed.  Remove it from
the general ``workspace_rsync_excludes`` list.

Signed-off-by: Chris Evich <cevich@redhat.com>